test(#zimic): split large body test files by type
